requirements.txt

asgiref
certifi
charset-normalizer
Django

exceptiongroup
gunicorn
h11
httpcore
httpx
idna
packaging
psycopg2-binary
python-dotenv
python-telegram-bot[rate-limiter]
python-dateutil
pytz
requests
six
sniffio
sqlparse
tzlocal
urllib3
sentry-sdk

# REST Framework & API
djangorestframework
django-filter
django-cors-headers

# Celery & Redis
django-celery-beat
django-celery-results
celery
redis

# JWT Authentication & Database
python-jose[cryptography]
dj-database-url

# Static files
whitenoise

# Document processing
lxml==5.4.0
python-docx==1.2.0
typing_extensions==4.14.0

# AI
mistralai

feedparser
careerjet-api-client